requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
streamlit>=1.20.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
pytest>=7.4.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
pypdf>=3.0.0
python-multipart>=0.0.6
groq==0.31.1
